BASELIST 3.0
UPDATE :  re added missing query to clear the value if user changes action

UPDATE:  Now when deal is selected validation required is added to the  initial fee field. If user adds a value then changes the action the value in the initial field is cleared and the required validation is also removed.