BASELIST 3.0 -
New interface element introduced into the add address pop up. If the trading address is the same as the registered address a user can now click on the newly created checkbox and import the registered address into the text field.

To compliment this functionality. On the add company page similar  functionality has been added to cater for all scenarios.

On the company detail page within the email engagement section I noticed  the algorithm was slightly off target which resulted in no company user email events being presented on some companies. This has now been fixed and has been successfully tested with 100% accuracy.

Finally the blank page that sometimes appear when a user logs in has been sorted. The current solution is not suitable for the long term and I will need to change asap.