BASELIST 3.0
Apostrophes in autocomplete - try to search for Kiri O'Brien in live
Issue: I found a range of issues that mainly related to input integrity.

1) When  contact e.g Paddy O'Reilly  was inserted in the search field a 400 error was thrown on the server.
2) If an apostrophe was added to a user name the raw character would be inserted into the database. I have changed this to instead insert a safe html entity.

The above issues had also be resolved in the contacts and addresses controllers/model.

Email Setting header changed from Settings to Email Settings

Contact page  -  Changed to match the look and feel of the rest of the page