Baselist 3.10
Fixed issue relating to favourites on the company details page.

Issue: When user clicked on star button the action to show or hide user name pillar next to pipeline if set was not evaluating.

Problem: Header class evaluation error was caused by the introduction of the breadcrumb parent header tag being introduced to facilitate the introduction of a sticky header.