Baselist 3.12
Without notice I found out that 4 days ago Companies House changed the address variable [snippet] to [address_snippet]. As a  resulted when a company was created via the API the address would not pull in the data.  Instead the word unknown would populate the database. Issue reported by RL this morning and issue found, fix and pushed to production as I considered this a critical issue.