Lacer Unicorns Sprint 3 - ZD fix
Resolved non showing of on hold data.

Resolved skewed total ticket qty being presented. Now only evaluating ticket status only an nothing else in the returned JSON count.

In creased iteration of all tickets to 500. Within the loop if we encounter a ticket with a new or open status we show the first five regardless of there status.

Fixed visibility on header values. If a status of new or open is registered then visibility will be set to show. Default visibility is hide;